+++ lib/src/core/models/code/code_entities/field.dart
@@ -5,8 +5,10 @@ class Field extends CodeEntity {
   final Map<String, Value> values;
 
   String enumName(String theme) {
-    return '${path.skip(1).map((e) => e.upperFirst).join()}${theme != Config.it.baseTheme ? theme.upperFirst : ''}'
-        .lowerFirst;
+    return [
+      ...path.skip(1),
+      theme != Config.it.baseTheme ? theme : '',
+    ].joinCamelCase.lowerFirst;
   }
 
   Field({