+++ lib/src/core/parsers/file_parsers/json_file_parser.dart
@@ -5,72 +5,72 @@ import 'package:genuis/src/core/models/nodes/node.dart';
 import 'package:genuis/src/core/parsers/file_parsers/file_parser.dart';
 import 'package:genuis/src/utils/exceptions.dart';
 import 'package:genuis/src/utils/file_system_entity_extension.dart';
-import 'package:genuis/src/utils/num_extension.dart';
-import 'package:genuis/src/utils/string_extension.dart';
+
+typedef _JsonMap = Map<String, dynamic>;
+typedef _JsonList = List<dynamic>;
 
 class JsonFileParser extends FileParser {
   @override
   bool canParse(File file) => file.isJson;
 
   @override
-  List<Node> parse(File file, String name) {
-    if (!file.isJson) return [];
+  Node parse(File file) {
+    if (!file.isJson) return Folder.empty;
 
     try {
-      final root = json.decode(file.readAsStringSync());
+      final content = json.decode(file.readAsStringSync());
 
-      return _parseJson({name: root});
+      return _parseEntry(content);
     } catch (e) {
       if (e is ParserFileElementException) throw ParserFileException(file.path, element: e.element);
       throw ParserFileException(file.path);
     }
   }
 
-  List<Node> _parseJson(Map<String, dynamic> json) {
+  Node _parseEntry(dynamic value) {
+    if (value is _JsonMap) {
+      return Folder(
+        name: '',
+        nodes: _parseJson(value),
+      );
+    }
+
+    if (value is _JsonList) {
+      return Folder(
+        name: '',
+        nodes: value.map((e) => _parseEntry(e)).toList(),
+      );
+    }
+
+    if (value is String || value is int || value is double || value is bool) {
+      return Item(
+        value: value.toString(),
+      );
+    }
+
+    throw ParserFileElementException('{$value}');
+  }
+
+  List<Node> _parseJson(_JsonMap json) {
     List<Node> nodes = [];
 
     for (final key in json.keys) {
       final entity = json[key];
-      if (entity is Map<String, dynamic>) {
-        nodes.add(Folder(
-          name: key.camelCase.named,
-          nodes: _parseJson(entity),
-        ));
-        continue;
-      }
-      if (entity is List) {
-        nodes.addAll(
-          entity.map(
-            (e) {
-              var name = key;
-              if (e is double) {
-                if (e.isInt) {
-                  name += e.toStringAsFixed(0);
-                } else {
-                  name += e.toString().snakeCase.named;
-                }
-              } else {
-                name += e.toString().camelCase.named;
-              }
-
-              return Item(
-                name: name,
-                value: e.toString(),
-              );
-            },
+      if (entity is _JsonMap) {
+        nodes.add(
+          Folder(
+            name: key,
+            nodes: _parseJson(entity),
           ),
         );
-        continue;
+      } else {
+        nodes.add(
+          Folder(
+            name: key,
+            nodes: [_parseEntry(entity)],
+          ),
+        );
       }
-      if (entity is String || entity is int || entity is double) {
-        nodes.add(Item(
-          name: key.camelCase.named,
-          value: entity.toString(),
-        ));
-        continue;
-      }
-
-      throw ParserFileElementException('{$key: $entity}');
     }
 
     return nodes;