+++ lib/src/utils/file_system_entity_extension.dart
@@ -3,14 +3,15 @@ import 'dart:io';
 import 'package:path/path.dart';
 import 'package:genuis/src/utils/string_extension.dart';
 
-extension DirectoryExtension on Directory {
-  String get name => path.forwardSlash.split('/').lastWhere((e) => e.isNotEmpty);
+extension FileSystemEntityExtension on FileSystemEntity {
+  String get name {
+    final name = basename(path);
+    if (this is File) return name.withoutExtension;
+    return name;
+  }
 }
 
 extension FileExtension on File {
-  String get name => path.forwardSlash.split('/').lastWhere((e) => e.isNotEmpty).withoutExtension;
-
   bool get isXml => extension(path) == '.xml';
-
   bool get isJson => extension(path) == '.json';
 }