+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Fix: issue with collecting themes
+* Fix: ignoring separators in json-keys
+
 ## 0.0.1
 
 * Initial development release.